pyproject.toml
[project]
name = "k8s-mcp"
version = "0.1.0"
description = "A read-only Model Context Protocol (MCP) for querying Kubernetes clusters."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kubernetes>=32.0.1",
    "mcp[cli]>=1.3.0",
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.9",
]
